requirements.txt
caffe2==0.8.1
docutils==0.21.2
einops==0.8.0
GPUtil==1.4.0
GPUtil==1.4.0
matplotlib==3.10.0
ml_collections==0.1.1
numpy==2.2.1
packaging==24.2
Pillow==11.0.0
psutil==6.0.0
pynvml==12.0.0
scipy==1.14.1
setuptools==75.1.0
Sphinx==8.1.3
sphinx_rtd_theme==3.0.2
tabulate==0.9.0
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
triton==3.0.0
xmlrunner==1.7.7
